% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-base-class.R
\name{BaseComponent}
\alias{BaseComponent}
\alias{InterceptComponent}
\title{Class for defining component interface to CIDnetwork object}
\description{
An object used internally by the CIDnetwork object, which interfaces with
each component in the model via a standardized set of functions. This class
allows developers to create plug-and-play model components
by cutomizing the expected functions. Each component has a \code{Component} class
that inherits from the \code{BaseComponent} class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseComponent$new()}}
\item \href{#method-random.start}{\code{BaseComponent$random.start()}}
\item \href{#method-draw}{\code{BaseComponent$draw()}}
\item \href{#method-create.output.list}{\code{BaseComponent$create.output.list()}}
\item \href{#method-update.output.list}{\code{BaseComponent$update.output.list()}}
\item \href{#method-get.mean.component}{\code{BaseComponent$get.mean.component()}}
\item \href{#method-value}{\code{BaseComponent$value()}}
\item \href{#method-clone}{\code{BaseComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$new(edge.list, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random.start"></a>}}
\if{latex}{\out{\hypertarget{method-random.start}{}}}
\subsection{Method \code{random.start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$random.start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-draw"></a>}}
\if{latex}{\out{\hypertarget{method-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$draw(outcome, residual.variance)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create.output.list"></a>}}
\if{latex}{\out{\hypertarget{method-create.output.list}{}}}
\subsection{Method \code{create.output.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$create.output.list(total.draws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update.output.list"></a>}}
\if{latex}{\out{\hypertarget{method-update.output.list}{}}}
\subsection{Method \code{update.output.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$update.output.list(gibbs.output.list, draw)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.mean.component"></a>}}
\if{latex}{\out{\hypertarget{method-get.mean.component}{}}}
\subsection{Method \code{get.mean.component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$get.mean.component(gibbs.output.list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
