% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-base-class.R
\name{BaseParams}
\alias{BaseParams}
\alias{InterceptParams}
\title{Class for storing and transporting component parameters}
\description{
An object used internally by the CIDnetwork object when initializing the
components of the model. Returned by user-facing component functions like
\code{INTERCEPT}, which allow the user to define prior parameters and intial values
for the Gibbs sampler. Each component has a \code{Params} class that inherits
from the \code{BaseParams} class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseParams$new()}}
\item \href{#method-create.component}{\code{BaseParams$create.component()}}
\item \href{#method-clone}{\code{BaseParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseParams$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create.component"></a>}}
\if{latex}{\out{\hypertarget{method-create.component}{}}}
\subsection{Method \code{create.component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseParams$create.component(n.nodes, edge.list, node.names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.nodes}}{}

\item{\code{edge.list}}{}

\item{\code{node.names}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
