\name{CIDnetworks-data}
\alias{Lazega}
\alias{c.elegans}
\alias{dolphins}

\title{CIDnetworks: Included Data Sets}
\description{
  Data sets included with the CIDnetworks package.
}
\usage{
  Lazega
  c.elegans
  dolphins
}
\format{
  Lazega: A list including six sociomatrices (three directed, three
  symmetrized) and one attributes matrix for the nodes.

  c.elegans: An symmetric valued ``sociomatrix'' for the number of
  connections between neurons.

  dolphins: A symmetric sociomatrix.
}
\references{
  
  Emmanuel Lazega. The Collegial Phenomenon: The Social Mechanisms of
  Cooperation Among Peers in a Corporate Law Partnership, Oxford
  University Press (2001).

  D. J. Watts and S. H. Strogatz, Nature 393, 440-442 (1998).

  J. G. White, E. Southgate, J. N. Thompson, and S. Brenner, Phil. Trans. R. Soc. London 314, 1-340 (1986).
  
  D. Lusseau, K. Schneider, O. J. Boisseau, P. Haase, E. Slooten, and
  S. M. Dawson, Behavioral Ecology and Sociobiology 54, 396-405 (2003).
 
}
\author{A.C. Thomas <act@acthomas.ca>}
\keyword{datasets}
