% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-intercept.R
\name{INTERCEPT}
\alias{INTERCEPT}
\title{Specify intercept component for a CIDnetwork model}
\usage{
INTERCEPT(intercept.m = 0, intercept.v = 1000, intercept = 0)
}
\arguments{
\item{intercept.m}{Numeric scalar specifying the prior mean (default = 0).}

\item{intercept.v}{Numeric scalar specifying the prior variance (must be > 0;
default = 1000).}

\item{intercept}{Numeric scalar defining the initial value for sampling
(default = 0).}
}
\value{
An \code{InterceptParams} object that can be fed in a list of other
components to the \code{components} argument of \code{CIDnetwork$new()}.
}
\description{
Specify the prior mean and variance and starting value for the intercept
component of your CIDnetwork model. Returns an \code{InterceptParams} object that
stores the your supplied arguments. This output should be passed as part of a
\code{list} of other components to the \code{components} argument of the
\code{CIDnetwork$new()} initialization method to define your model.
}
