% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-SBM.R
\name{SBM}
\alias{SBM}
\title{Specify SBM component for a CIDnetwork model}
\usage{
SBM(
  n.groups = 2,
  block.matrix.m = matrix(0, nrow = n.groups, ncol = n.groups),
  block.matrix.v = matrix(10000, nrow = n.groups, ncol = n.groups),
  membership.a = NULL,
  symmetric.b = TRUE,
  strong.block = FALSE,
  block.matrix = NULL,
  membership = NULL
)
}
\arguments{
\item{n.groups}{Integer scalar specifying the number of groups/blocks.
(default = 2)}

\item{block.matrix.m}{Numeric matrix of size \verb{G X G} specifying the mean of
the beta prior distribution for each block-block tie probability. \code{G} is
the number of groups/blocks. (default = matrix of all 0s)}

\item{block.matrix.v}{Numeric matrix of size \verb{G X G} specifying the variance
of the beta prior distribution for each block-block tie probability. \code{G} is
the number of groups/blocks. (default = matrix of all 1000s; "weak" prior,
nearly uniform)}

\item{membership.a}{Numeric matrix of size \verb{N X G} specifying the prior
probability for each node to be a member of each group. \code{N} is the number
of nodes and \code{G} is the number of groups/blocks.  Rows do not need to sum
to 1, as MCMC takes care of normalization. However, it is highly
recommended to specify rows that sum to 1 if not using a uniform
probability (like the default) so that you are fully aware of the
normalized prior probability you are specifying. (default = matrix of all
1s)}

\item{symmetric.b}{Logical scalar, if \code{TRUE}, will force \code{block.matrix} to be
symmetric. Ensures that group sending-receiving probabilities are
identical. (default = TRUE)}

\item{block.matrix}{Numeric matrix of size \verb{G X G} specifying the initial
block-block tie probability matrix. \code{G} is the number of groups/blocks.
(default = NULL, which randomly generates the matrix from the prior)}

\item{membership}{Numeric vector of length \code{N} specifying the initial value
for the group/block membership vector. \code{N} is the number of nodes. (default
= NULL, which randomly generates the vector from the prior)}

\item{stong.block}{Logical scalar, if \code{TRUE}, will force \code{block.matrix} to
have larger diagonal entries than off-diagonal. Ensures that intra-group
tie probabilities are strong than ties with nodes in other groups. (default
= FALSE)}
}
\value{
An \code{SBMParams} object that can be fed in a list of other components
to the \code{components} argument of \code{CIDnetwork$new()}.
}
\description{
Specify the prior mean and variance for the block-block tie matrix, the
multinomial prior on the membership vector, and their starting values for the
SBM component of your CIDnetwork model. Returns an \code{SBMParams} object that
stores the your supplied arguments. This output should be passed as part of a
\code{list} of other components to the \code{components} argument of the
\code{CIDnetwork$new()} initialization method to define your model.
}
\details{
In a stochastic block model (SBM), the probability of observing a network tie
between two nodes depends on which groups (called blocks) the nodes are
members of, and the relationship between the groups. Nodes in the same block
are more likely to have a tie than nodes in different block; and some times
certain pairs of blocks are more likely to have nodes that share ties than
other pairs.

The model is written in notation as
\deqn{Y_{ij} \sim Bernoulli(g_i^T B g_j)}
\deqn{g_i \sim Multinomial(p_i)}
\deqn{B_{\ell m} \sim Beta(\mu_{\ell m}, \nu_{\ell m})}

This package uses the mean-variance parameterization to specify the prior on
the block-block tie matrix.

The variables in this model correspond to the following arguments for this
function
\itemize{
\item \eqn{g_i} is the group/block membership vector. You can pass the
initial value in the \code{membership} argument.
\item \eqn{p_i} is the prior for the group/block membership vector which is
the \code{membership.a} argument. In this package, the prior can be specified
for each node.
\item \eqn{B_{\ell m}} is the block-block tie matrix. You can pass the
initial value in the \code{block.matrix} argument.
\item \eqn{\mu_{\ell m}} corresponds to the \code{block.matrix.m} argument.
\item \eqn{\nu_{\ell m}} corresponds to the \code{block.matrix.v} argument.
}

The block.matrix priors can be used to incorporate information about how
likely inter-group ties are.
}
