\name{CIDnetworks-master}
\alias{CID}
\alias{CID.generate}
\alias{CIDnetwork}

\title{Forward Simulation and Declaration of CID Network Models}
\description{
  Usage of the CID main class.
}
\usage{

  CID (...)
  CID.generate (...)

}
\arguments{

  \item{...}{Arguments passed to the master class.}
    
}
\details{
  Generating a complex network according to the CID modelling framework
  is easy and straightforward. See vignettes for more information.  
}
\value{
  A CIDnetwork object.
}
\author{A.C. Thomas <act@acthomas.ca>}

